+++ crates/tsz-checker/src/state/state_checking/mod.rs
@@ -8,6 +8,7 @@ mod core;
 pub(crate) mod directive;
 mod dts_rules;
 pub(crate) mod heritage;
+mod heritage_call_constructor;
 mod heritage_class_recovery;
 mod heritage_support;
 mod isolated_declarations;
